sttt/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai-whisper==20250625
torch>=2.2.0
torchaudio>=2.2.0
librosa==0.10.1
soundfile==0.12.1
numpy>=1.21.0
